fixed a bug in v*printf() where '%%' was not handled correctly in all cases
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@6119 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/QoreObject.cpp
@@ -1221,7 +1221,7 @@ int ObjectRSetHelper::checkIntern(QoreObject& obj) {
       fi->second = true;
       printd(QRO_LVL, "ObjectRSetHelper::checkIntern() done search obj %p '%s': no match, removing from current set\n", &obj, obj.getClassName());
    }
-   else if (mrc == ovec.size()) {
+   else if (mrc == (int)ovec.size()) {
       assert(frvec.size() > fpos);
 
       // we have found a complete independent recursive cycle at this level, so apply it and mark nodes as finalized
